feat(blockchain): centralize types and update status
