+++ docs/wiki/development/component-status.md
@@ -54,13 +54,13 @@ This document provides a comprehensive test status report for all components in 
 | @smolitux/ai | SentimentDisplay | ✅ | ❌ | ❌ | ❌ | Needs A11y Tests |
 | @smolitux/ai | TrendingTopics | ✅ | ❌ | ❌ | ❌ | Needs A11y Tests |
 | @smolitux/ai | TrollFilter | ✅ | ❌ | ❌ | ❌ | Needs A11y Tests |
-| @smolitux/blockchain | SmartContractInteraction | ✅ | ❌ | ❌ | ❌ | Needs A11y Tests |
-| @smolitux/blockchain | StakingInterface | ✅ | ❌ | ❌ | ❌ | Needs A11y Tests |
-| @smolitux/blockchain | TokenDisplay | ✅ | ❌ | ❌ | ❌ | Needs A11y Tests |
-| @smolitux/blockchain | TokenDistributionChart | ✅ | ❌ | ❌ | ❌ | Needs A11y Tests |
-| @smolitux/blockchain | TokenEconomy | ✅ | ❌ | ❌ | ❌ | Needs A11y Tests |
-| @smolitux/blockchain | TransactionHistory | ✅ | ❌ | ❌ | ❌ | Needs A11y Tests |
-| @smolitux/blockchain | WalletConnect | ✅ | ❌ | ❌ | ❌ | Needs A11y Tests |
+| @smolitux/blockchain | SmartContractInteraction | ✅ | ✅ | ✅ | ✅ | Ready |
+| @smolitux/blockchain | StakingInterface | ✅ | ✅ | ✅ | ✅ | Ready |
+| @smolitux/blockchain | TokenDisplay | ✅ | ✅ | ✅ | ✅ | Ready |
+| @smolitux/blockchain | TokenDistributionChart | ✅ | ✅ | ✅ | ✅ | Ready |
+| @smolitux/blockchain | TokenEconomy | ✅ | ✅ | ✅ | ✅ | Ready |
+| @smolitux/blockchain | TransactionHistory | ✅ | ✅ | ✅ | ✅ | Ready |
+| @smolitux/blockchain | WalletConnect | ✅ | ✅ | ✅ | ✅ | Ready |
 | @smolitux/community | ActivityFeed | ✅ | ❌ | ❌ | ❌ | Needs A11y Tests |
 | @smolitux/community | CommentSection | ✅ | ❌ | ❌ | ❌ | Needs A11y Tests |
 | @smolitux/community | FollowButton | ✅ | ❌ | ❌ | ❌ | Needs A11y Tests |
@@ -115,7 +115,7 @@ This document provides a comprehensive test status report for all components in 
 - **@smolitux/layout**: 2 components ready, 2 need A11y tests
 - **@smolitux/charts**: 7 components ready
 - **@smolitux/ai**: 8 components need A11y tests
-- **@smolitux/blockchain**: 7 components need A11y tests
+- **@smolitux/blockchain**: All components complete with A11y tests
 - **@smolitux/community**: 5 components need A11y tests
 - **@smolitux/federation**: 5 components need A11y tests
 - **@smolitux/media**: 6 components ready, 1 needs A11y tests
@@ -161,13 +161,13 @@ Dieser Bericht wird mit jeder Version aktualisiert, um den Fortschritt bei der T
 | @smolitux/ai | SentimentDisplay | ⚠️ Teilweise |
 | @smolitux/ai | TrendingTopics | ⚠️ Teilweise |
 | @smolitux/ai | TrollFilter | ⚠️ Teilweise |
-| @smolitux/blockchain | SmartContractInteraction | ⚠️ Teilweise |
-| @smolitux/blockchain | StakingInterface | ⚠️ Teilweise |
-| @smolitux/blockchain | TokenDisplay | ⚠️ Teilweise |
-| @smolitux/blockchain | TokenDistributionChart | ⚠️ Teilweise |
-| @smolitux/blockchain | TokenEconomy | ⚠️ Teilweise |
-| @smolitux/blockchain | TransactionHistory | ⚠️ Teilweise |
-| @smolitux/blockchain | WalletConnect | ⚠️ Teilweise |
+| @smolitux/blockchain | SmartContractInteraction | ✅ Fertig |
+| @smolitux/blockchain | StakingInterface | ✅ Fertig |
+| @smolitux/blockchain | TokenDisplay | ✅ Fertig |
+| @smolitux/blockchain | TokenDistributionChart | ✅ Fertig |
+| @smolitux/blockchain | TokenEconomy | ✅ Fertig |
+| @smolitux/blockchain | TransactionHistory | ✅ Fertig |
+| @smolitux/blockchain | WalletConnect | ✅ Fertig |
 | @smolitux/charts | AreaChart | ✅ Fertig |
 | @smolitux/charts | BarChart | ✅ Fertig |
 | @smolitux/charts | Heatmap | ⚠️ Teilweise |
