--- packages/@smolitux/blockchain/src/components/TokenEconomy/TokenEconomy.test.tsx
@@ -1,21 +1,0 @@
-import React from 'react';
-import { render, screen } from '@testing-library/react';
-import { TokenEconomy } from './TokenEconomy';
-
-describe('TokenEconomy', () => {
-  it('renders without crashing', () => {
-    render(<TokenEconomy />);
-    expect(screen.getByRole('button', { name: /TokenEconomy/i })).toBeInTheDocument();
-  });
-
-  it('applies custom className', () => {
-    render(<TokenEconomy className="custom-class" />);
-    expect(screen.getByRole('button')).toHaveClass('custom-class');
-  });
-
-  it('forwards ref correctly', () => {
-    const ref = React.createRef<HTMLButtonElement>();
-    render(<TokenEconomy ref={ref} />);
-    expect(ref.current).toBeInstanceOf(HTMLButtonElement);
-  });
-});
